changing constructor to use np.int64 when parsing node feats

+++ graph_constructor.py
@@ -454,7 +454,7 @@ class GraphConstructor:
             'edge_index': np.array([[edge[0] for edge in edges_reindexed], [edge[1] for edge in edges_reindexed]]),
             'edge_feat': np.array([self.ONEHOT_EDGETYPE[edge[2]] for edge in edges_reindexed]),
             'node_feat': np.array([[float(x) for x in list(graph_only_refs[key].values())[2:]]
-            for key in graph_only_refs.keys()]),
+            for key in graph_only_refs.keys()]).astype(np.int64),
             'num_nodes': num_nodes,
             },
             output
